% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_action}
\alias{add_action}
\title{Add an action}
\usage{
add_action(x, ...)
}
\arguments{
\item{x}{The object to add the action to.}

\item{...}{Additional arguments (enables methods to use them).}
}
\value{
See specific implementations.
}
\description{
Add an action to an object (generic method)
}
