% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{iterate}
\alias{iterate}
\title{Iterate (generic)}
\usage{
iterate()
}
\arguments{
\item{x}{The thing to iterate from/on}

\item{...}{Additional arguments.}
}
\value{
See specific implementations.
}
\description{
Iterate from an object
}
